+++ src/java/dawson/songtracker/types/Components/SongComponent.java
@@ -10,9 +10,9 @@ public abstract class SongComponent {
     protected final int id;
     protected String name;
     protected final Timestamp creationTime;
-    protected int duration;
+    protected double duration;
 
-    public SongComponent(int id, String name, Timestamp creationTime, int duration) {
+    public SongComponent(int id, String name, Timestamp creationTime, double duration) {
         if(name == null){
             throw new NullPointerException("the name is null");
         }
@@ -42,16 +42,16 @@ public abstract class SongComponent {
         return creationTime;
     }
 
-    public int getDuration() {
+    public double getDuration() {
         return duration;
     }
 
     public String getDurationString() {
 //        Duration of 1hour 2minutes 47seconds
 //        duration = 3767
-        int hours = (duration - (duration % 3600)) / 3600;
-        int seconds = (duration - hours * 3600) % 60;
-        int minutes = (duration - hours * 3600 - seconds) / 60;
+        int hours = (int) (duration - (duration % 3600)) / 3600;
+        int seconds = (int) (duration - hours * 3600) % 60;
+        int minutes = (int) (duration - hours * 3600 - seconds) / 60;
         return (hours < 10 ? "0" + hours : hours) + ":" + (minutes < 10 ? "0" + minutes : minutes) + "."
                 + (seconds < 10 ? "0" + seconds : seconds);
     }
